+++ apps/cli/CHANGELOG.md
@@ -1,5 +1,12 @@
 # walkeros
 
+## 3.3.1
+
+### Patch Changes
+
+- Updated dependencies [62f6a38]
+  - @walkeros/cli@3.3.1
+
 ## 3.3.0
 
 ### Patch Changes