+++ apps/walkerjs/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @walkeros/walker.js
 
+## 3.3.1
+
+### Patch Changes
+
+- Updated dependencies [b10144a]
+- Updated dependencies [206185a]
+- Updated dependencies [50e5d09]
+- Updated dependencies [32ff626]
+  - @walkeros/collector@3.3.1
+  - @walkeros/web-source-browser@3.3.1
+  - @walkeros/web-source-datalayer@3.3.1
+  - @walkeros/web-source-session@3.3.1
+  - @walkeros/core@3.3.1
+  - @walkeros/web-core@3.3.1
+
 ## 3.3.0
 
 ### Patch Changes