+++ packages/server/destinations/tiktok/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @walkeros/server-destination-meta
 
+## 3.3.1
+
+### Patch Changes
+
+- @walkeros/server-core@3.3.1
+- @walkeros/core@3.3.1
+
 ## 3.3.0
 
 ### Patch Changes