+++ packages/server/transformers/file/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @walkeros/server-transformer-file
 
+## 3.3.1
+
+### Patch Changes
+
+- @walkeros/core@3.3.1
+
 ## 3.3.0
 
 ### Patch Changes